refactor: MVC패턴 적용
