+++ src/main/java/baseball/model/User.java
@@ -1,7 +1,6 @@
-package baseball.domain;
+package baseball.model;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
 public class User {
